fix: moved home.jsc to its own folder inside src
